we import the euro value and then convert to cents import csv corected and add expense done part 1

+++ app/Domain/Entity/Expense.php
@@ -13,7 +13,7 @@ final class Expense
         public int $userId,
         public DateTimeImmutable $date,
         public string $category,
-        public float $amountCents,
+        public int $amountCents,
         public string $description,
     ) {}
 }
